MAINT: Apply ruff rule RUF046
RUF046 Value being cast to `int` is already an integer

+++ asv/util.py
@@ -116,7 +116,7 @@ def human_file_size(size, err=None):
     if size == 0:
         num_scale = 0
     else:
-        num_scale = int(math.floor(math.log(size) / math.log(1000)))
+        num_scale = math.floor(math.log(size) / math.log(1000))
     if num_scale > 7:
         suffix = '?'
     else:
